# Added _enabledControllers cache to OpenSim::Model