This PR is to try and optimize certain (mostly, controller-heavy) simulations, such as the FD sim for `ToyDropLanding`. In those simulations, profilers (`valgrind`/`perf`) reported that ~8 % of CPU time was spent calling `getComponentList`.

The reason that `getComponentList` can have this kind of overhead is because its implementation involves `dynamic_cast`ing (downcasting) every element in a container to determine whether the element is a `Controller` or not (the container's elements are `Component`s). The `computeControls` method can be called many times in a simulation, so this iterate-and-downcast can happen *a lot*.

The proposed fix for this is to cache a `std::vector<Controller const&>` of enabled `Controller`s in the `Model` (`Model::_enabledControllers`) just after the model's connection graph is fully established. At this point in `Model`'s lifecycle, `Controller`s and other subcomponents shouldn't change, so the cache should remain valid during the remainder of the simulation. The cache is used by `computeControls` at runtime.

Overall, this caching change makes *some* simulations 5-10 % faster. Usually the simulations containing a bunch of muscles. Not a huge performance improvement--something I'm desperately searching for in OpenSim ;)--but not a terrible performance improvement for a relatively small change (albeit, a change that increases the complexity of `Model`). 

Overall stats from my current measurement technique, which effectively just runs some FD/IK simulations (the suite is work in progress, and should probably be ran on different hardware, etc. temporary repo [here](https://github.com/adamkewley/throwaway-osim-perf)).

```
source activate && ./comparison
```

repeats = 64
|                  Test Name | master [secs] | σ [secs] | branch [secs] | σ [secs] | Speedup |
| -------------------------- | ------------- | -------- | ------------- | -------- | ------- |
|                   Gait2354 |          0.27 |     0.00 |          0.27 |     0.00 |    0.99 |
|             ToyDropLanding |         15.41 |     0.05 |         14.43 |     0.06 |    1.07 |
|   ToyDropLanding_nomuscles |          0.45 |     0.00 |          0.45 |     0.00 |    1.00 |
|            passive_dynamic |          5.10 |     0.02 |          5.07 |     0.01 |    1.01 |
| passive_dynamic_noanalysis |          3.20 |     0.01 |          3.22 |     0.01 |    1.00 |
|                      Arm26 |          0.31 |     0.00 |          0.29 |     0.00 |    1.07 |
|             RajagopalModel |         15.88 |     0.06 |         15.10 |     0.06 |    1.05 |

---

Commit note:

- Caches all enabled OpenSim::Controllers in a model after
  the model graph is finalized (pre-computation)

- At compute time, used by Model::computeControllers as a
  faster substitute to using getComponentList<T>

- This is purely a performance modification. Benchmarks
  (e.g. on ToyDropLanding, from the OpenSim Tutorials)
  indicate a 5-10 % speedup with this change

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/opensim-org/opensim-core/2806)
<!-- Reviewable:end -->