@adamkewley 2020-06-17
So there's CI failures due to not being able to upload to sourceforge or something (ignoring this), but the errors with the Windows build are because MSVC (effectively) thinks this is invalid C++:

```c++
if (not x) {
    // bleh
}
```

Because it thinks `not` is an identifier, rather than a keyword. `not` is a standards-compliant C++ keyword, so it's unusual for the compiler to reject it (MSFT has it's reasons, I imagine). The solution is to either remove the `not`s (eugh) or add `/permissive-` to MSVC's build flags:

https://docs.microsoft.com/en-us/cpp/build/reference/permissive-standards-conformance?view=vs-2019

I'll try and add it into `.travis` etc. when I get a second.