@adamkewley 2020-06-17
Great review @chrisdembia . I'll make the changes tomorrow and push them up. They're all reasonable. 

For the `not` thing, it's purely something I've picked up from the C++ codebases I worked on previously. I became a fan because it can make for nice-to-read when combined in some awkward flow combined:

```c++
if (user.logged_in() and user.is_admin()) {
    if (not user.can_delete(other_user)) {
        return ErrorReponse{"ruh roh"};
    }
}
```

I didn't really think about the implicit `std::reference_wrapper<T>` -> `T&` conversion also - good catch.