@adamkewley 2020-06-18
This looks like it's good to go, is the main outstanding concern the removal of the enablement logic?