@chrisdembia 2020-06-18
> is the main outstanding concern the removal of the enablement logic?

Yes, can you address that concern?

The Travis CI tests are failing.

The CHANGELOG should also be updated, especially if the behavior of that function is changing.