@adamkewley 2020-06-19
The `CHANGELOG.md` has been updated. 

The reason those travis tests are failing is because of a sourceforge encryption key, not the build or the tests. According to the logs, everything is building + passing. The failure is happening during some `ssh` into sourceforge (presumably, to deploy binaries?).