@aseth1 2020-06-19
>is the main outstanding concern the removal of the enablement logic?

Looks to me that the logic was not removed but put in Model to decide whether to override the individual Controllers and disable all Controllers at once. This is effectively what was happening before, but the individual Controller had to look up into it Model, which was confusing. Now it is centralized and the Model is the final arbiter right before the System is created. IMHO the fix is preferred.