@chrisdembia 2020-06-19
Hey @aseth1; I was wondering if other parts of the code base were depending on `Controller::isEnabled()` or if `Model::computeControls()` contains the only call site.