@chrisdembia 2020-06-22
> So, we're good to go?

No; I clarified my question above:

> Hey @aseth1; I was wondering if other parts of the code base were depending on Controller::isEnabled() or if Model::computeControls() contains the only call site.

I've gone through the code myself and found two other uses of `Controller::isEnabled()`:

https://github.com/opensim-org/opensim-core/blob/c00135abd92d4d7ac155d1e604dfa1612493ea3c/OpenSim/Simulation/Model/ControllerSet.cpp#L81-L91

https://github.com/opensim-org/opensim-core/blob/c00135abd92d4d7ac155d1e604dfa1612493ea3c/OpenSim/Simulation/Model/ControllerSet.cpp#L97-L102

The behavior at these two locations is now different. I suggest the `Model::getAllControllersEnabled()` be employed here.