@aseth1 2020-06-22
@chrisdembia thanks for looking into the other places where `Controller::getEnabled()` is used. IMHO, what we had was incorrect, because `Model::_allControllersEnabled` is not a property and can be used (by design I believe) to override the computing of controls by controllers at System execution time.  Your suggestion would not enable a Controller to be "toggled" during computations since it would never be a part of the connected Model. Now a Controller is wired up only if it is individually enabled. At compute time, the Model can still override. I think this is more clear and less error prone behavior.